WIP: fsms: use configurable timers instead of T23042
It looks like T23042 was a placeholder for timers to be filled in later,
it cannot be configured via VTY and I cannot find it mentioned in any
specs on etsi.org. git blame shows it first appeard in 31f525e ("large
refactoring: use FSMs for lchans; add inter-BSC HO"). Replace T23042
with timers that can be configured via VTY.

Previous timeout for the states was 5s.

* ASSIGNMENT_ST_WAIT_MGW_ENDPOINT_TO_MSC and
  HO_ST_WAIT_MGW_ENDPOINT_TO_MSC:
  * Runs gscon_connect_mgw_to_msc() on enter, which waits for one MGCP
    CRCX or MDCX response (or changes state immediatelly)
  * Use existing X9 ("Timeout for availability of MGW endpoint"), 5s,
    which was previously being used by lchan_rtp_fsm for a similar
    purpuse

* HO_ST_WAIT_LCHAN_ACTIVE:
  * Runs lchan_activate() on enter, and waits for the lchan fsm to go
    through LCHAN_ST_WAIT_TS_READY (X5) and LCHAN_ST_WAIT_ACTIV_ACK
    (X6).
  * Use new X29, 10s (5s from X5 + 5s from X6)

Change-Id: Id0d4d0788f609f3272fc81c80a754383dde25c16